refactor(scraper): minor code cleanup as suggested in code review
- replace () => {return {...};} with () => ({...})
- remove commented-out code
- move utils to their own files in /app/utils
  - create '#utils/' subpath imports for /app/utils
